requirements.txt
fastapi==0.111.0
pydantic-settings==2.3.3

SQLAlchemy==2.0.31
psycopg2==2.9.9

PyJWT==2.8.0
bcrypt==4.1.3

redis==5.0.6
openai==1.35.3

validate-docbr==1.10.0
cachetools==5.3.3